Guards only see what is in front of them
NB: this is a wee bit broken, because guards currently scan and then
move.

The guard direction variable is set via their move function (which is
called after they have already finished scanning). So guards scan based
on the direction they were moving LAST turn, not the direction they are
moving THIS turn.

This can be fixed by changing the order in which guards process their
turn. They need to move (based on the scan they undertook the previous
turn), and then scan.